fix: reoganize safe pipes
