+++ apps/gauzy/src/app/@shared/pipes/index.ts
@@ -1,12 +1,12 @@
 import { DurationFormatPipe } from './duration-format.pipe';
 import { FilterArrayPipe } from './filter-array.pipe';
-import { Nl2BrPipe, SafeHtmlPipe, TruncatePipe } from './text.pipe';
+import { Nl2BrPipe, TruncatePipe } from './text.pipe';
 import { ReplacePipe } from './replace.pipe';
 import { TimeFormatPipe } from './time-format.pipe';
 import { UtcToLocalPipe } from './utc-to-local.pipe';
 import { DateTimeFormatPipe } from './datetime-format.pipe';
 import { DateFormatPipe } from './date-format.pipe';
-import { SafePipe } from './safe.pipe';
+import { SafeHtmlPipe, SafePipe } from './safe/safe.pipe';
 
 export * from './duration-format.pipe';
 export * from './filter-array.pipe';
@@ -16,7 +16,7 @@ export * from './time-format.pipe';
 export * from './datetime-format.pipe';
 export * from './date-format.pipe';
 export * from './utc-to-local.pipe';
-export * from './safe.pipe';
+export * from './safe/safe.pipe';
 
 export const Pipes = [
 	DurationFormatPipe,
